move buffer traits to a separate filemake
